tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/historical.html

<!doctype html>
<meta charset=utf-8>
<title>Historical canvas features</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
var canvas, context;
setup(function() {
  canvas = document.createElement("canvas");
  context = canvas.getContext('2d');
});
function t(member, obj) {
  var name = obj === canvas ? "Canvas" : String(obj).match(/\[object (\S+)\]/)[1];
  test(function() {
    assert_false(member in obj);
  }, name + " support for " + member);
}
// added in https://github.com/whatwg/html/commit/0ecbf0e010df16d9c6d11eef6b2c58419158c4da
// renamed in https://github.com/whatwg/html/commit/2542a12cb25ee93534cbed1f31b5e1bc05fcdd0e
t("supportsContext", canvas);

// removed in https://github.com/whatwg/html/commit/2cfb8e3f03d3166842d2ad0f661459d26e2a40eb
t("probablySupportsContext", canvas);

// removed in https://github.com/whatwg/html/commit/ef72f55da4acdf266174225c6ca8bf2a650d0219
t("width", context);
t("height", context);

// removed in https://github.com/whatwg/html/commit/740634d0f30a3b76e9da166ac2fa8835fcc073ab
t("setContext", canvas);
t("transferControlToProxy", canvas);
t("CanvasProxy", window);
t("commit", canvas);
test(function() {
  assert_throws(new TypeError(), function() {
    new CanvasRenderingContext2D();
  }, 'no arguments');
  assert_throws(new TypeError(), function() {
    new CanvasRenderingContext2D(1, 1);
  }, 'with arguments');
}, "CanvasRenderingContext2D constructors");
</script>
